requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openai==1.35.10
pandas==2.2.2
pyaes==1.6.1
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
Telethon==1.36.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
